+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-# Changelog for 1.3.1
-- Initial release of the mod to the public
+# Changelog for 1.3.2
+- First public release of the mod (for real this time, 1.3.1 didn't make it).
+- More sensible default configuration, now by default not changing the health or food level a player respawns with nor giving a death position map.
+- Small clarifications to option descriptions in the configuration screen.